Consolidated logic for all period date based rules

+++ test/ApplicationERC20Handler.t.sol
@@ -40,7 +40,7 @@ contract ApplicationERC20HandlerTest is Test, DiamondTestUtil, RuleProcessorDiam
     address ac;
     address[] badBoys;
     address[] goodBoys;
-    uint256 Blocktime = 1675723152;
+    uint64 Blocktime = 1675723152;
     RuleProcessorDiamond ruleProcessor;
     ApplicationHandler public applicationHandler;
     RuleStorageDiamond ruleStorageDiamond;
